TEST: moves playback tests to appropriate file
