+++ tests/basic_usage_test.rs
@@ -1,4 +1,4 @@
-use kenku_control::{*, playlist::playback};
+use kenku_control::*;
 use rand::Rng;
 
 const DEFAULT_IP_ADDRESS: &str = "127.0.0.1";
@@ -21,19 +21,4 @@ async fn play_a_randon_track() {
         None => ()
     }
 }
-    
-#[tokio::test]
-async fn pause_playlist_playback() {
-    let controller = Controller::new(DEFAULT_IP_ADDRESS, DEFAULT_PORT);
-    let command = playback::playback_pause(&controller).await.unwrap();
 
-    assert_eq!(command.is_success(), true);
-}
-
-#[tokio::test]
-async fn play_playlist_playback() {
-    let controller = Controller::new(DEFAULT_IP_ADDRESS, DEFAULT_PORT);
-    let command = playback::playback_play(&controller).await.unwrap();
-
-    assert_eq!(command.is_success(), true);
-}
